+++ src/views/Pools/components/LockedPool/utils/convertLockTimeToSeconds.ts
@@ -1,0 +1,3 @@
+export default function convertLockTimeToSeconds(lockTime: string): number {
+  return parseInt(lockTime) * 1000
+}
